Applications appendix rewrite
Removed most of the text i considered non-essential or focusing on something other than applications.
Also removed most of the ending about how metadata affects topics as we already do mention it in the article and the questions we ask are not answered anywhere so they might end up hurting us.
The new text is focused more on giving more examples of applications and making the examples a bit more concrete.